+++ data/src/main/java/com/thk/data/database/SevenDaysDao.kt
@@ -13,16 +13,16 @@ interface ChallengeDao {
     fun getChallenges(): Flow<List<Challenge>>
 
     @Insert(onConflict = IGNORE)
-    suspend fun addChallenge(challenge: Challenge): Int
+    suspend fun addChallenge(challenge: Challenge): Long
 
     @Query("DELETE FROM ${DatabaseInfo.TABLE_NAME_CHALLENGE} WHERE ${DatabaseInfo.COLUMN_NAME_CHALLENGE_ID} = :challengeId")
-    suspend fun removeChallenge(challengeId: Int)
+    suspend fun removeChallenge(challengeId: Long)
 }
 
 @Dao
 interface StampDao {
     @Query("SELECT * FROM ${DatabaseInfo.TABLE_NAME_STAMP} WHERE ${DatabaseInfo.COLUMN_NAME_CHALLENGE_ID} = :challengeId")
-    suspend fun getStamps(challengeId: Int): List<Stamp>
+    suspend fun getStamps(challengeId: Long): List<Stamp>
 
     @Insert(onConflict = REPLACE)
     suspend fun addStamps(vararg stamp: Stamp): List<Int>
